# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Inbox Server",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/audio-uploads": {
      "post": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Create Audio Upload",
        "operationId": "create_audio_upload_v1_audio_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioUploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUploadSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio-uploads/{session_id}": {
      "get": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Get Audio Upload",
        "operationId": "get_audio_upload_v1_audio_uploads__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUploadSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Cancel Audio Upload",
        "description": "取消未完成会话释放临时占用；已完成对象按引用规则处理，不删共享对象。",
        "operationId": "cancel_audio_upload_v1_audio_uploads__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Audio Upload V1 Audio Uploads  Session Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio-uploads/{session_id}/chunks": {
      "put": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Put Audio Chunk",
        "description": "一次一个块；同一 offset 同摘要重传幂等，不同摘要冲突，不接受跳段/重叠。",
        "operationId": "put_audio_chunk_v1_audio_uploads__session_id__chunks_put",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUploadSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio-uploads/{session_id}/complete": {
      "post": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Complete Audio Upload",
        "description": "长度一致后分块计算整文件 SHA-256，并把 staging 原子收纳为不可变对象。",
        "operationId": "complete_audio_upload_v1_audio_uploads__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/items/{item_id}/audio-original": {
      "get": {
        "tags": [
          "audio-uploads"
        ],
        "summary": "Download Audio Original",
        "description": "鉴权流式下载上传原件：只按登记的 AudioAsset 定位，不暴露服务器路径。",
        "operationId": "download_audio_original_v1_items__item_id__audio_original_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioUploadCreate": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "total_bytes": {
            "type": "integer",
            "title": "Total Bytes"
          },
          "mime": {
            "type": "string",
            "title": "Mime",
            "default": "application/octet-stream"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "total_bytes"
        ],
        "title": "AudioUploadCreate"
      },
      "AudioUploadResult": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "sha256",
          "bytes",
          "mime",
          "filename"
        ],
        "title": "AudioUploadResult"
      },
      "AudioUploadSessionOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "total_bytes": {
            "type": "integer",
            "title": "Total Bytes"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size"
          },
          "upload_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "max_duration_seconds": {
            "type": "integer",
            "title": "Max Duration Seconds"
          },
          "max_audio_bytes": {
            "type": "integer",
            "title": "Max Audio Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "state",
          "filename",
          "mime",
          "total_bytes",
          "offset",
          "chunk_size",
          "upload_id",
          "expires_at",
          "max_duration_seconds",
          "max_audio_bytes"
        ],
        "title": "AudioUploadSessionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://kb.jerrythealpaca.cn"
    }
  ]
}